+++ data/create-org.ts
@@ -1,0 +1,27 @@
+const sizeOptions = [
+  { label: "1-10", value: "1-10" },
+  { label: "11-50", value: "11-50" },
+  { label: "51-100", value: "51-100" },
+  { label: "101-500", value: "101-500" },
+  { label: "501-1000", value: "501-1000" },
+  { label: "1000+", value: "1000+" },
+];
+
+const focusOptions = [
+  { label: "Religion", value: "religion" },
+  { label: "Software", value: "software" },
+  { label: "Engineering", value: "engineering" },
+  { label: "Health Tech", value: "healthTech" },
+  { label: "Crypto", value: "crypto" },
+];
+
+const roleOptions = [
+  { label: "CEO/Owner", value: "ceo/owner" },
+  { label: "HR Manager", value: "hrManager" },
+  { label: "IT/Tech Manager", value: "it/techManager" },
+  { label: "IT/Tech Staff", value: "it/techStaff" },
+  { label: "HR Staff", value: "hrStaff" },
+  { label: "Other", value: "other" },
+];
+
+export const formOptions = {sizeOptions, focusOptions, roleOptions};
